코드 정리
함수로 분할

기존 controlButton 모양이 안맞던 거 수정

파일 명 수정, extension 기능  재구분
