+++ WOD/Timer/Common/View/Time/ControlButton.swift
@@ -8,31 +8,33 @@
 import SwiftUI
 
 struct ControlButton: View {
-    @State private var isPaused: Bool = false
+    @Binding var isPaused: Bool
+
     var action: (() -> Void)?
     let img1Name: String?
     let img2Name: String?
     let defaultImgName: String?
     
-    init(action: (@escaping () -> Void), img1Name: String?, img2Name: String?, defaultImgName: String? = nil) {
+    init(isPaused: Binding<Bool>, action: (@escaping () -> Void), img1Name: String?, img2Name: String?, defaultImgName: String? = nil) {
+        self._isPaused = isPaused
         self.action = action
         self.img1Name = img1Name
         self.img2Name = img2Name
         self.defaultImgName = nil
     }
     
-    init(action: (@escaping () -> Void), img1Name: String? = nil, img2Name: String? = nil, defaultImgName: String) {
+    init(isPaused: Binding<Bool>, action: (@escaping () -> Void), img1Name: String? = nil, img2Name: String? = nil, defaultImgName: String) {
+        self._isPaused = isPaused
         self.action = action
         self.img1Name = nil
         self.img2Name = nil
         self.defaultImgName = defaultImgName
     }
-        
+    
     var body: some View {
         GeometryReader { geometry in
             ZStack {
                 Button(action: {
-                    isPaused.toggle()
                     guard let action = action else { return }
                     action()
                 }) {
